Split out proj4def and proxy service to separate Express routers.
