+++ server/crs.js
@@ -1,0 +1,22 @@
+"use strict";
+var express = require('express');
+var router = express.Router();
+
+var proj4 = require('proj4');
+
+//TODO: check if this loads the file into each core and if so then,
+require('proj4js-defs/epsg')(proj4);
+
+
+//provide REST service for proj4 definition strings
+router.get('/:crs', function(req, res, next) {
+    var crs = req.param('crs');
+    var epsg = proj4.defs[crs.toUpperCase()];
+    if (epsg !== undefined) {
+        res.status(200).send(epsg);
+    } else {
+        res.status(500).send('no proj4 definition');
+    }
+});
+
+module.exports = router;
