+++ server/index.js
@@ -1,168 +1,11 @@
 "use strict";
 
 
-var url = require('url');
-var configSettings = require('../wwwroot/config.json');
-var proxyAuth = require('../proxyAuth.json');
+var proxy = require('./proxy');
+var crs = require('./crs');
 
-var protocolRegex = /^\w+:\//;
-
-var upstreamProxy;
-var bypassUpstreamProxyHosts;
-
-var dontProxyHeaderRegex = /^(?:Host|X-Forwarded-Host|Proxy-Connection|Connection|Keep-Alive|Transfer-Encoding|TE|Trailer|Proxy-Authorization|Proxy-Authenticate|Upgrade)$/i;
-
-function filterHeaders(req, headers) {
-    var result = {};
-    // filter out headers that are listed in the regex above
-    Object.keys(headers).forEach(function(name) {
-        if (!dontProxyHeaderRegex.test(name)) {
-            result[name] = headers[name];
-        }
-    });
-
-    return result;
-}
-
-function filterResponseHeaders(req, headers, maxAgeSeconds) {
-    var result = filterHeaders(req, headers);
-
-    result['Cache-Control'] = 'public,max-age=' + maxAgeSeconds;
-    result['Access-Control-Allow-Origin'] ='*';
-    delete result['Expires'];
-    delete result['pragma'];
-
-    return result;
-}
-
-var proxyDomains = configSettings.proxyDomains;
-
-
-//Non CORS hosts and domains we proxy to
-function proxyAllowedHost(host) {
-    host = host.toLowerCase();
-    //check that host is from one of these domains
-    for (var i = 0; i < proxyDomains.length; i++) {
-        if (host.indexOf(proxyDomains[i], host.length - proxyDomains[i].length) !== -1) {
-            return true;
-        }
-    }
-    return false;
-}
-
-var durationRegex = /^([\d.]+)(ms|s|m|h|d|w|y)$/;
-
-var durationUnits = {
-    ms: 1.0 / 1000,
-    s: 1.0,
-    m: 60.0,
-    h: 60.0 * 60.0,
-    d: 24.0 * 60.0 * 60.0,
-    w: 7.0 * 24.0 * 60.0 * 60.0,
-    y: 365.0 * 24.0 * 60.0 * 60.0
-};
-
-function doProxy(req, res, next, callback) {
-    var maxAgeSeconds = 1209600; // two weeks
-    var remoteUrlString = req.params[0];
-
-    if (!remoteUrlString || remoteUrlString.length === 0) {
-        return res.status(400).send('No url specified.');
-    }
-
-    // Does the proxy URL include a max age?
-    if (remoteUrlString[0] === '_') {
-        var slashIndex = remoteUrlString.indexOf('/');
-        if (slashIndex < 0) {
-            return res.status(400).send('No url specified.');
-        }
-
-        var maxAgeString = remoteUrlString.substring(1, slashIndex);
-        remoteUrlString = remoteUrlString.substring(slashIndex + 1);
-
-        if (remoteUrlString.length === 0) {
-            return res.status(400).send('No url specified.');
-        }
-
-        // Interpret the max age as a duration in Varnish notation.
-        // https://www.varnish-cache.org/docs/trunk/reference/vcl.html#durations
-        var parsedMaxAge = durationRegex.exec(maxAgeString);
-        if (!parsedMaxAge || parsedMaxAge.length < 3) {
-            return res.status(400).send('Invalid duration.');
-        }
-
-        var value = parseFloat(parsedMaxAge[1]);
-        if (value !== value) {
-            return res.status(400).send('Invalid duration.');
-        }
-
-        var unitConversion = durationUnits[parsedMaxAge[2]];
-        if (!unitConversion) {
-            return res.status(400).send('Invalid duration unit ' + parsedMaxAge[2]);
-        }
-
-        maxAgeSeconds = value * unitConversion;
-    }
-
-    // Add http:// if no protocol is specified.
-    var protocolMatch = protocolRegex.exec(remoteUrlString);
-    if (!protocolMatch || protocolMatch.length < 1) {
-        remoteUrlString = 'http://' + remoteUrlString;
-    } else {
-        var matchedPart = protocolMatch[0];
-
-        // If the protocol portion of the URL only has a single slash after it, the extra slash was probably stripped off by someone
-        // along the way (NGINX will do this).  Add it back.
-        if (remoteUrlString[matchedPart.length] !== '/') {
-            remoteUrlString = matchedPart + '/' + remoteUrlString.substring(matchedPart.length);
-        }
-    }
-
-    var remoteUrl = url.parse(remoteUrlString);
-
-    // Copy the query string
-    remoteUrl.search = url.parse(req.url).search;
-
-    if (!remoteUrl.protocol) {
-        remoteUrl.protocol = 'http:';
-    }
-
-    var proxy;
-    if (upstreamProxy && !(remoteUrl.host in bypassUpstreamProxyHosts)) {
-        proxy = upstreamProxy;
-    }
-
-    // Are we allowed to proxy for this host?
-    if (!proxyAllowedHost(remoteUrl.host)) {
-        res.status(400).send('Host is not in list of allowed hosts: ' + remoteUrl.host);
-        return;
-    }
-
-    // encoding : null means "body" passed to the callback will be raw bytes
-
-    var proxiedRequest;
-    req.on('close', function() {
-        if (proxiedRequest) {
-            proxiedRequest.abort();
-        }
-    });
-
-    var filteredReqHeaders = filterHeaders(req, req.headers);
-    if (!filteredReqHeaders['x-forwarded-for']) {
-        filteredReqHeaders['x-forwarded-for'] = req.connection.remoteAddress;
-    }
-
-    // http basic auth
-    var authRequired = proxyAuth[remoteUrl.host];
-    if (authRequired) {
-        filteredReqHeaders['authorization'] = authRequired.authorization;
-    }
-
-    proxiedRequest = callback(remoteUrl, filteredReqHeaders, proxy, maxAgeSeconds);
-}
-
-// Include the cluster module
 var cluster = require('cluster');
+var request = require('request');
 
 // Code to run if we're in the master process
 if (cluster.isMaster) {
@@ -197,15 +40,10 @@ if (cluster.isMaster) {
     var express = require('express');
     var fs = require('fs');
     var compression = require('compression');
-    var request = require('request');
     var path = require('path');
     var cors = require('cors');
     var formidable = require('formidable');
     var ogr2ogr = require('ogr2ogr');
-    var proj4 = require('proj4');
-
-    //TODO: check if this loads the file into each core and if so then,
-    require('proj4js-defs/epsg')(proj4);
 
     var yargs = require('yargs').options({
         'port' : {
@@ -249,11 +87,12 @@ if (cluster.isMaster) {
     app.disable('etag');
     app.use(express.static(path.join(__dirname, 'wwwroot')));
 
-    upstreamProxy = argv['upstream-proxy'];
-    bypassUpstreamProxyHosts = {};
+    var po = proxy._proxyOptions = {};
+    po.upstreamProxy = argv['upstream-proxy'];
+    po.bypassUpstreamProxyHosts = {};
     if (argv['bypass-upstream-proxy-hosts']) {
         argv['bypass-upstream-proxy-hosts'].split(',').forEach(function(host) {
-            bypassUpstreamProxyHosts[host.toLowerCase()] = true;
+            po.bypassUpstreamProxyHosts[host.toLowerCase()] = true;
         });
     }
 
@@ -261,56 +100,9 @@ if (cluster.isMaster) {
       res.status(200).send('OK');
     });
 
-    app.get('/proxy/*', function(req, res, next) {
-        doProxy(req, res, next, function(remoteUrl, filteredRequestHeaders, proxy, maxAgeSeconds) {
-            return request.get({
-                url : url.format(remoteUrl),
-                headers : filteredRequestHeaders,
-                encoding : null,
-                proxy : proxy
-            }, function(error, response, body) {
-                var code = 500;
+    app.use('/proxy', proxy);
 
-                if (response) {
-                    code = response.statusCode;
-                    res.header(filterResponseHeaders(req, response.headers, maxAgeSeconds));
-                }
-
-                res.status(code).send(body);
-            });
-        });
-    });
-
-    app.post('/proxy/*', function(req, res, next) {
-        doProxy(req, res, next, function(remoteUrl, filteredRequestHeaders, proxy, maxAgeSeconds) {
-            req.pipe(request.post({
-                url : url.format(remoteUrl),
-                headers : filteredRequestHeaders,
-                encoding : null,
-                proxy : proxy
-            }, function(error, response, body) {
-                var code = 500;
-
-                if (response) {
-                    code = response.statusCode;
-                    res.header(filterResponseHeaders(req, response.headers, maxAgeSeconds));
-                }
-
-                res.status(code).send(body);
-            }));
-        });
-    });
-
-    //provide REST service for proj4 definition strings
-    app.get('/proj4def/:crs', function(req, res, next) {
-        var crs = req.param('crs');
-        var epsg = proj4.defs[crs.toUpperCase()];
-        if (epsg !== undefined) {
-            res.status(200).send(epsg);
-        } else {
-            res.status(500).send('no proj4 definition');
-        }
-    });
+    app.use('/proj4def', crs);
 
     // provide conversion to geojson service
     // reguires install of gdal on server: sudo apt-get install gdal-bin
